Move LogCell code into a folder
